Revert "add: API通信時のローディング画面へのページ遷移作成"
