+++ nativeApp/src/navigation/stacks/HomeStack.js
@@ -4,7 +4,6 @@ import PictureScreen from "../../screens/PictureScreen";
 import { createStackNavigator } from "@react-navigation/stack";
 import { CameraProvider } from "../../context/CameraContext";
 import GetMaterialScreen from "../../screens/GetMaterialScreen";
-import SubmitToAPIScreen from "../../screens/SubmitToAPIScreen";
 
 const Stack = createStackNavigator();
 
@@ -15,7 +14,6 @@ const HomeScreenNavigator = () => {
         <Stack.Screen name="Home" component={HomeScreen} />
         <Stack.Screen name="Camera" component={CameraScreen} />
         <Stack.Screen name="Picture" component={PictureScreen} />
-        <Stack.Screen name="SubmitToAPI" component={SubmitToAPIScreen} />
         <Stack.Screen
           name="GetMaterial"
           component={GetMaterialScreen}
