+++ nativeApp/src/screens/PictureScreen.js
@@ -1,18 +1,36 @@
+import React from "react";
 import { StyleSheet, View, Image, Dimensions } from "react-native";
 import { useCamera } from "../context/CameraContext";
 import ChangeMaterialButton from "../components/ChangeMaterialButton";
+import sendImage from "../lib/sendPicture";
+import { selectData } from "../lib/dataBaseHelper";
+import { RO, Material } from "../lib/databaseQueryText";
+import getRandomNum from "../lib/getRandomNum";
 
 const windowWidth = Dimensions.get("window").width;
 const windowHeight = Dimensions.get("window").height;
 
 const PictureScreen = ({ navigation }) => {
   const { picture } = useCamera();
-  console.log(picture)
-  
 
-  const handleSubmitToAPI = () => {
-    navigation.navigate("SubmitToAPI", {
-      picture:picture,
+  const handleSubmitToAPI = async () => {
+    const response = await sendImage({ uri: picture.uri });
+    selectData(
+      Material.tablename,
+      Material.elementsKey.colorId,
+      RO.Eqqual,
+      response.return
+    ).then((materials) => {
+      if (materials.length > 0) {
+        const randomIndx = getRandomNum(materials.length);
+        navigation.navigate("GetMaterial", {
+          getMaterial: materials[randomIndx],
+        });
+      } else {
+        console.log(
+          `colorId == ${response.return}のマテリアルが登録されていません`
+        );
+      }
     });
   };
   return (
